+++ src/library/DIPS.Mobile.UI/Components/Buttons/iOS/ButtonHandler.cs
@@ -21,7 +21,7 @@ public partial class ButtonHandler : Microsoft.Maui.Handlers.ButtonHandler
 
     private static partial void MapImageToRightSide(ButtonHandler handler, Button button)
     {
-        if(!button.ImageToRightSide)
+        if(button.ImagePlacement == ImagePlacement.Left)
             return;
         
         handler.PlatformView.Transform = CGAffineTransform.MakeScale(-1, 1);